Quetoo Forge Game Data
Removed the model file paths from the FGD. Jackhammer handles
this
weird. Where the entity would not show up in the editor. Where
as
leaving filepath blank results in an entity showing. Will revise
this
later.

Also need to adjust the entity sizes as some are flat as
pancakes on
creation. However, the FGD is fully compatible as it stands
with
JACKHAMMER and Trenchbroom. Worked with all the *.map files I've
tried
it with included in the respositories.