Trenchbroom 2.0.0.0 Support
Added this folder for easy drag and drop access to map in Quetoo in
Trenchbroom 2.0.0.0  ... The Quetoo.cfg needs to be revised, but was put
together for quick access into the editor. Be warned with using this
setup you will get an error or two. And working on texture support as
well. Just things to keep in mind while mapping in Trenchbroom. Will be
revised sometime soon.